Rename file for useActiveBackend hook
